backend/requirements-fastapi.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.22
authlib>=1.4.0
httpx>=0.28.0
itsdangerous>=2.2.0
python-dotenv>=1.1.0
psutil>=7.0.0
aiofiles>=24.0.0
python-multipart>=0.0.18
python-socketio>=5.13.0
pydantic-settings>=2.7.0
